Αλλαξα ωστε τα δυο μεγαλυτερα στοιχεια να μην ειναι ιδια

+++ kand7/Askisi1/askisi1.h
@@ -1,6 +1,6 @@
 // askisi1.h
 using namespace std;
 
-void twoMaxer(double array[], double size);
+void twoMaxer(double array[], int size);
 double &smallest(double array[], double size);
-double *creatArray(double size);
+double *creatArray(int size);
